refactor code (resolve typo, move component atomically)
